builder-util-runtime: separate newError to eliminate circular dependency
