+++ packages/builder-util-runtime/src/error.ts
@@ -1,0 +1,5 @@
+export function newError(message: string, code: string) {
+  const error = new Error(message)
+  ;(error as NodeJS.ErrnoException).code = code
+  return error
+}
